move some specs out of active_model_spec
