+++ spec/unit/query_proxy_methods_spec.rb
@@ -1,0 +1,164 @@
+require 'spec_helper'
+
+describe 'query_proxy_methods' do
+  # goofy names to differentiate from same classes used elsewhere
+  before(:all) do
+    class IncludeLesson; end
+    class IncludeTeacher; end
+    class IncludeEmptyClass; end
+    class IncludeStudent
+      include Neo4j::ActiveNode
+      property :name
+      has_many :out, :lessons, model_class: IncludeLesson, type: 'lessons'
+      has_many :out, :things, model_class: false, type: 'lessons'
+    end
+
+    class IncludeLesson
+      include Neo4j::ActiveNode
+      property :name
+      has_many :in, :students, model_class: IncludeStudent, origin: :lessons
+      has_many :in, :teachers, model_class: IncludeTeacher, origin: :lessons
+    end
+
+    class IncludeTeacher
+      include Neo4j::ActiveNode
+      has_many :out, :lessons, model_class: IncludeLesson, type: 'teaching_lesson'
+    end
+
+    class IncludeEmptyClass
+      include Neo4j::ActiveNode
+      has_many :out, :lessons, model_class: IncludeLesson
+    end
+  end
+  let!(:jimmy)    { IncludeStudent.create(name: 'Jimmy') }
+  let!(:math)     { IncludeLesson.create(name: 'math') }
+  let!(:science)  { IncludeLesson.create(name: 'science') }
+  let!(:mr_jones) { IncludeTeacher.create }
+  let!(:mr_adams) { IncludeTeacher.create }
+
+  describe 'first and last' do
+    it 'returns objects across multiple associations' do
+      jimmy.lessons << science
+      science.teachers << mr_adams
+      expect(jimmy.lessons.teachers.first).to eq mr_adams
+      expect(mr_adams.lessons.students.last).to eq jimmy
+    end
+  end
+
+  describe 'include?' do
+    it 'correctly reports when a node is included in a query result' do
+      jimmy.lessons << science
+      science.teachers << mr_adams
+      expect(jimmy.lessons.include?(science)).to be_truthy
+      expect(jimmy.lessons.include?(math)).to be_falsey
+      expect(jimmy.lessons.teachers.include?(mr_jones)).to be_falsey
+      expect(jimmy.lessons.where(name: 'science').teachers.include?(mr_jones)).to be_falsey
+      expect(jimmy.lessons.where(name: 'science').teachers.include?(mr_adams)).to be_truthy
+      expect(IncludeTeacher.all.include?(mr_jones)).to be_truthy
+      expect(IncludeTeacher.all.include?(math)).to be_falsey
+    end
+
+    it 'works with multiple relationships to the same object' do
+      jimmy.lessons << science
+      jimmy.lessons << science
+      expect(jimmy.lessons.include?(science)).to be_truthy
+    end
+
+    it 'returns correctly when model_class is false' do
+      woodworking = IncludeLesson.create(name: 'woodworking')
+      expect(jimmy.things.include?(woodworking)).to be_falsey
+      jimmy.lessons << woodworking
+      expect(jimmy.things.include?(woodworking)).to be_truthy
+      woodworking.destroy
+    end
+
+    it 'allows you to check for an identifier in the middle of a chain' do
+      jimmy.lessons << science
+      science.teachers << mr_adams
+      expect(IncludeLesson.as(:l).students.where(name: 'Jimmy').include?(science, :l)).to be_truthy
+    end
+
+    it 'raises an error if something other than a node is given' do
+      expect { IncludeStudent.lessons.include?(:foo) }.to raise_error(Neo4j::ActiveNode::Query::QueryProxyMethods::InvalidParameterError)
+    end
+  end
+
+  describe 'exists?' do
+    context 'class methods' do
+      it 'can run by a class' do
+        expect(IncludeEmptyClass.empty?).to be_truthy
+        expect(IncludeLesson.empty?).to be_falsey
+      end
+
+      it 'can be called with a property and value' do
+        expect(IncludeLesson.exists?(name: 'math')).to be_truthy
+        expect(IncludeLesson.exists?(name: 'boat repair')).to be_falsey
+      end
+
+      it 'can be called on the class with a neo_id' do
+        expect(IncludeLesson.exists?(math.neo_id)).to be_truthy
+        expect(IncludeLesson.exists?(8675309)).to be_falsey
+      end
+
+      it 'raises an error if something other than a neo id is given' do
+        expect { IncludeLesson.exists?(:fooooo) }.to raise_error(Neo4j::ActiveNode::QueryMethods::InvalidParameterError)
+      end
+    end
+
+    context 'QueryProxy methods' do
+      it 'can be called on a query' do
+        expect(IncludeLesson.where(name: 'history').exists?).to be_falsey
+        expect(IncludeLesson.where(name: 'math').exists?).to be_truthy
+      end
+
+      it 'can be called with property and value' do
+        expect(jimmy.lessons.exists?(name: 'science')).to be_falsey
+        jimmy.lessons << science
+        expect(jimmy.lessons.exists?(name: 'science')).to be_truthy
+        expect(jimmy.lessons.exists?(name: 'bomb disarming')).to be_falsey
+      end
+
+      it 'can be called with a neo_id' do
+        expect(IncludeLesson.where(name: 'math').exists?(math.neo_id)).to be_truthy
+        expect(IncludeLesson.where(name: 'math').exists?(science.neo_id)).to be_falsey
+      end
+
+      it 'is called by :blank? and :empty?' do
+        expect(jimmy.lessons.blank?).to be_truthy
+        expect(jimmy.lessons.empty?).to be_truthy
+        jimmy.lessons << science
+        expect(jimmy.lessons.blank?).to be_falsey
+        expect(jimmy.lessons.empty?).to be_falsey
+      end
+    end
+  end
+
+  describe 'count' do
+    before(:all) do
+      @john = IncludeStudent.create(name: 'Paul')
+      @history = IncludeLesson.create(name: 'history')
+      3.times { @john.lessons << @history }
+    end
+
+    it 'tells you the number of matching objects' do
+      expect(@john.lessons.count).to eq(3)
+    end
+
+    it 'can tell you the number of distinct matching objects' do
+      expect(@john.lessons.count(:distinct)).to eq 1
+    end
+
+    it 'raises an exception if a bad parameter is passed' do
+      expect { @john.lessons.count(:foo) }.to raise_error(Neo4j::ActiveNode::Query::QueryProxyMethods::InvalidParameterError)
+    end
+
+    it 'works on an object earlier in the chain' do
+      expect(IncludeStudent.as(:s).lessons.where(name: 'history').count(:distinct, :s)).to eq 1
+    end
+
+    it 'is aliased by length and size' do
+      expect(@john.lessons.size).to eq(3)
+      expect(@john.lessons.length).to eq(3)
+    end
+  end
+end
